Move `StripUnsignedFromEventsTestCase` to a more specific file
